+++ sdk/python/pvsite_datamodel/write/forecast.py
@@ -4,62 +4,18 @@ Functions for writing to pvsite db
 
 import datetime as dt
 import logging
-import typing
 import uuid
 
 import dateutil.parser as dp
 import numpy as np
 import pandas as pd
-import sqlalchemy as sa
-import sqlalchemy.dialects.postgresql as sa_psql
 import sqlalchemy.orm as sa_orm
 from pvsite_datamodel import  sqlmodels
 
 # Defines the length of time over which a forecast is valid
-FORECAST_TIMESPAN = dt.timedelta(minutes=5)
-
-
-class WrittenRow(typing.NamedTuple):
-    """
-    Defines a write to the Database
-    """
-
-    table: sqlmodels.Base
-    pk_value: str
-
-
-def upsert(session: sa_orm.Session, table: sqlmodels.Base, rows: list[dict]) -> list[WrittenRow]:
-    """
-    Upsert rows into table
-
-    This functions checks the primary keys, and if present, updates the row.
-    :param session: sqlalchemy Session
-    :param table: the table
-    :param rows: the rows we are going to update
-    :return list[WrittenRow]: A list of WrittenRow objects containing the table and primary
-    key values
-    that have been written
-    """
-
-    # Input type validation in case user passes in a dict, not a list of dicts
-    if type(rows) != list:
-        if type(rows) == dict:
-            rows = [rows]
-        else:
-            raise TypeError("input rows must be a list of dict objects")
-
-    stmt = sa_psql.insert(table.__table__)
-    primary_key_names = [key.name for key in sa.inspect(table.__table__).primary_key]
-    update_dict = {c.name: c for c in stmt.excluded if not c.primary_key}
-
-    if not update_dict:
-        raise ValueError("insert_or_update resulted in an empty update_dict")
-
-    stmt = stmt.on_conflict_do_update(index_elements=primary_key_names, set_=update_dict)
-    session.execute(stmt, rows)
-    session.commit()
-
-    return [WrittenRow(table=table, pk_value=row[primary_key_names[0]]) for row in rows]
+from pvsite_datamodel.write.datetime_intervals import get_or_else_create_datetime_interval
+from pvsite_datamodel.write.upsert import upsert
+from pvsite_datamodel.write.utils import WrittenRow
 
 
 def insert_forecast_values(
@@ -146,42 +102,3 @@ def insert_forecast_values(
     return written_rows
 
 
-def get_or_else_create_datetime_interval(
-    session: sa_orm.Session, start_time: dt.datetime, end_time: dt.datetime = None
-) -> tuple[sqlmodels.DatetimeIntervalSQL, list[WrittenRow]]:
-    """
-    Gets a DatetimeInterval from the DB by start time if it exists, otherwise it creates a new entry
-
-    :param session: The SQLAlchemy session used for performing db updates
-    :param start_time: The start time of the datetime interval
-    :param end_time: The end time of the datetime interval. Optional, defaults to the start_time
-    + FORECAST_TIMESPAN
-    :return tuple(sqlmodels.DatetimeIntervalSQL, list[WrittenRow]): A tuple containing the existing
-    or created DatetimeIntervalSQL object, and a list of WrittenRow objects dictating what was
-    written to the DB
-    """
-
-    # End time defaults to the start time + FORECAST_TIMESPAN timedelta
-    if end_time is None:
-        end_time = start_time + FORECAST_TIMESPAN
-
-    # Check if a datetime interval exists for the input times
-    query = session.query(sqlmodels.DatetimeIntervalSQL)
-    query = query.filter(sqlmodels.DatetimeIntervalSQL.start_utc == start_time)
-    query = query.filter(sqlmodels.DatetimeIntervalSQL.end_utc == end_time)
-    existing_interval: sqlmodels.DatetimeIntervalSQL = query.first()
-
-    # If it does, fetch it's uuid
-    if existing_interval is not None:
-        return existing_interval, []
-
-    # If it doesn't, create a new one
-    else:
-        datetime_interval: sqlmodels.DatetimeIntervalSQL = sqlmodels.DatetimeIntervalSQL(
-            datetime_interval_uuid=uuid.uuid4(),
-            start_utc=start_time,
-            end_utc=end_time,
-            created_utc=dt.datetime.now(tz=dt.timezone.utc),
-        )
-        written_rows = upsert(session, sqlmodels.DatetimeIntervalSQL, [datetime_interval.__dict__])
-        return datetime_interval, written_rows
